+++ lang/src/org/partiql/lang/eval/EvaluatingCompiler.kt
@@ -439,16 +439,7 @@ internal class EvaluatingCompiler(
             is PartiqlAst.Expr.Bag -> compileSeq(ExprValueType.BAG, expr.values, metas)
 
             // bag operators
-            is PartiqlAst.Expr.BagOp -> {
-                err(
-                    "${expr.op.javaClass.canonicalName} is not yet supported",
-                    ErrorCode.EVALUATOR_FEATURE_NOT_SUPPORTED_YET,
-                    errorContextFrom(metas).also {
-                        it[Property.FEATURE_NAME] = expr.javaClass.canonicalName
-                    },
-                    internal = false
-                )
-            }
+            is PartiqlAst.Expr.BagOp -> compileBagOp(expr, metas)
         }
     }
 
@@ -518,8 +509,7 @@ internal class EvaluatingCompiler(
                 range.contains(longValue)
             }
             else -> error(
-                "The expression's static type was supposed to be INT but instead it was ${value.type}" +
-                    "This may indicate the presence of a bug in the type inferencer."
+                "IntegerValueValidator can only accept ExprValue with type INT, NULL, and MISSING."
             )
         }
     }
@@ -527,7 +517,7 @@ internal class EvaluatingCompiler(
     /**
      *  For operators which could return integer type, check integer overflow in case of [TypingMode.PERMISSIVE].
      */
-    private fun resolveIntConstraint(computeThunk: ThunkEnv, metas: MetaContainer): ThunkEnv =
+    private fun resolveArithmeticOverflow(computeThunk: ThunkEnv, metas: MetaContainer): ThunkEnv =
         when (val staticTypes = metas.staticType?.type?.getTypes()) {
             // No staticType, can't validate integer size.
             null -> computeThunk
@@ -547,53 +537,43 @@ internal class EvaluatingCompiler(
                             computeThunk
                         }
                     }
+
                     TypingMode.PERMISSIVE -> {
-                        // if null, then StaticTypes contains one or more Int Type.
-                        val nonIntType = staticTypes.filterNot {
-                            it is IntType
-                        }.ifEmpty { null }
                         val biggestIntegerType = staticTypes.filterIsInstance<IntType>().maxBy {
                             it.rangeConstraint.numBytes
                         }
-                        val validateResult = { validator: ((ExprValue) -> Boolean)?, naryResult: ExprValue ->
-                            errorSignaler.errorIf(
-                                !validator?.let { it(naryResult) }!!,
-                                ErrorCode.EVALUATOR_INTEGER_OVERFLOW,
-                                { ErrorDetails(metas, "Integer overflow", errorContextFrom(metas)) },
-                                { naryResult }
-                            )
-                        }
-                        when (nonIntType) {
-                            // only intType exists
-                            null -> {
-                                // biggestIntegerType can not be null in this case
-                                val validator = integerValueValidator(biggestIntegerType!!.rangeConstraint.validRange)
+                        when (biggestIntegerType) {
+                            // static type contains one or more IntType
+                            is IntType -> {
+                                val validator = integerValueValidator(biggestIntegerType.rangeConstraint.validRange)
                                 thunkFactory.thunkEnv(metas) { env ->
                                     val naryResult = computeThunk(env)
-                                    validateResult(validator, naryResult)
-                                }
-                            }
-                            // mix types, or only non intType
-                            else ->
-                                when (biggestIntegerType) {
-                                    // if mixed type, validate if naryResult is Int
-                                    // warning: intermediate result can potentially cause overflow. i.e., (MAX_INT2 + MAX_INT2) / 2.0
-                                    is IntType -> {
-                                        val validator = integerValueValidator(biggestIntegerType.rangeConstraint.validRange)
-                                        thunkFactory.thunkEnv(metas) { env ->
-                                            val naryResult = computeThunk(env)
-                                            when (naryResult.type) {
-                                                ExprValueType.INT -> {
-                                                    validateResult(validator, naryResult)
-                                                }
-                                                else -> naryResult
+                                    // validation shall only happen when the result is INT/MISS/NULL
+                                    // this is important as StaticType may contain a mixture of multiple types
+                                    when (val type = naryResult.type) {
+                                        ExprValueType.INT, ExprValueType.MISSING, ExprValueType.NULL -> errorSignaler.errorIf(
+                                            !validator(naryResult),
+                                            ErrorCode.EVALUATOR_INTEGER_OVERFLOW,
+                                            { ErrorDetails(metas, "Integer overflow", errorContextFrom(metas)) },
+                                            { naryResult }
+                                        )
+
+                                        else -> {
+                                            if (staticTypes.all { it is IntType }) {
+                                                error(
+                                                    "The expression's static type was supposed to be INT but instead it was $type" +
+                                                        "This may indicate the presence of a bug in the type inferencer."
+                                                )
+                                            } else {
+                                                naryResult
                                             }
                                         }
                                     }
-                                    // If only non-IntType StaticType, can't validate the integer size either.
-                                    null -> computeThunk
-                                    else -> computeThunk
                                 }
+                            }
+                            // If there is no IntType StaticType, can't validate the integer size either.
+                            null -> computeThunk
+                            else -> computeThunk
                         }
                     }
                 }
@@ -611,7 +591,7 @@ internal class EvaluatingCompiler(
             (lValue.numberValue() + rValue.numberValue()).exprValue()
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compileMinus(expr: PartiqlAst.Expr.Minus, metas: MetaContainer): ThunkEnv {
@@ -625,7 +605,7 @@ internal class EvaluatingCompiler(
             (lValue.numberValue() - rValue.numberValue()).exprValue()
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compilePos(expr: PartiqlAst.Expr.Pos, metas: MetaContainer): ThunkEnv {
@@ -638,7 +618,7 @@ internal class EvaluatingCompiler(
             value
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compileNeg(expr: PartiqlAst.Expr.Neg, metas: MetaContainer): ThunkEnv {
@@ -648,7 +628,7 @@ internal class EvaluatingCompiler(
             (-value.numberValue()).exprValue()
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compileTimes(expr: PartiqlAst.Expr.Times, metas: MetaContainer): ThunkEnv {
@@ -658,7 +638,7 @@ internal class EvaluatingCompiler(
             (lValue.numberValue() * rValue.numberValue()).exprValue()
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compileDivide(expr: PartiqlAst.Expr.Divide, metas: MetaContainer): ThunkEnv {
@@ -686,7 +666,7 @@ internal class EvaluatingCompiler(
             }
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compileModulo(expr: PartiqlAst.Expr.Modulo, metas: MetaContainer): ThunkEnv {
@@ -701,7 +681,7 @@ internal class EvaluatingCompiler(
             (lValue.numberValue() % denominator).exprValue()
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compileEq(expr: PartiqlAst.Expr.Eq, metas: MetaContainer): ThunkEnv {
@@ -1075,7 +1055,7 @@ internal class EvaluatingCompiler(
             }
         }
 
-        return resolveIntConstraint(computeThunk, metas)
+        return resolveArithmeticOverflow(computeThunk, metas)
     }
 
     private fun compileLit(expr: PartiqlAst.Expr.Lit, metas: MetaContainer): ThunkEnv {
@@ -1602,6 +1582,21 @@ internal class EvaluatingCompiler(
                 seqType,
                 makeItemThunkSequence(env)
             )
+        }
+    }
+
+    private fun compileBagOp(node: PartiqlAst.Expr.BagOp, metas: MetaContainer): ThunkEnv {
+        val lhs = compileAstExpr(node.operands[0])
+        val rhs = compileAstExpr(node.operands[1])
+        val op = ExprValueBagOp.create(node.op, metas)
+        return thunkFactory.thunkEnv(metas) { env ->
+            val l = lhs(env)
+            val r = rhs(env)
+            val result = when (node.quantifier) {
+                is PartiqlAst.SetQuantifier.All -> op.eval(l, r)
+                is PartiqlAst.SetQuantifier.Distinct -> op.eval(l, r).distinct()
+            }
+            valueFactory.newBag(result)
         }
     }